requirements.txt
pydantic
google-genai
groq
python-dotenv
openai
pyyaml
fastapi
uvicorn
gradio
requests